# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: theData.io API
  description: |
    theData.io API provides endpoints for event ingestion, analytics, and data management.
    
    ## Authentication
    All API endpoints require authentication using an API key that should be included
    in the `X-API-Key` header of each request.
    
    ## Rate Limiting
    API endpoints are rate-limited based on your subscription tier. Rate limit information
    is included in the response headers.
  version: '1.0.0'
  contact:
    name: theData.io Support
    url: https://thedata.io/support
    email: support@thedata.io
servers:
  - url: https://api.thedata.io/v1
    description: Production API
  - url: https://api.staging.thedata.io/v1
    description: Staging API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Event:
      type: object
      required:
        - event_type
        - timestamp
        - source
      properties:
        event_type:
          type: string
          description: Type of the event
        timestamp:
          type: string
          format: date-time
          description: Event timestamp
        source:
          type: string
          description: Source of the event
        user_id:
          type: string
          description: Associated user ID
        properties:
          type: object
          description: Custom event properties
        metadata:
          type: object
          description: Additional metadata
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
paths:
  /ingest/events:
    post:
      summary: Ingest a single event
      description: Send a single event for processing
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '202':
          description: Event accepted for processing
        '400':
          description: Invalid event data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
  /ingest/events/batch:
    post:
      summary: Ingest multiple events
      description: Send multiple events for batch processing
      security:
        - ApiKeyAuth: []
      parameters:
        - name: batch_size
          in: query
          schema:
            type: integer
            default: 100
          description: Size of each processing batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Event'
      responses:
        '202':
          description: Events accepted for processing
        '400':
          description: Invalid event data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
...